requirements.txt
hikari[speedups]==2.0.0.dev108
hikari-lightbulb==2.2.1
python-dotenv==0.20.0
pygelbooru==0.5.0
uvloop==0.16.0; os_name != "nt"